[FIX] sale_subscription: decrease dependance to deferred dates
Before this commit, the quantity invoiced in _prepare_invoice_line would
be compted based on the result of _get_subscription_qty_invoiced for a
future period. For product invoiced based on ordered quantity, it makes
no sense as the quantity is the ordered quantity. This commit split the
logic to reuse code when we need to know which quantity must be invoiced
or which quantity have been invoiced in the past.

closes odoo/enterprise#77912

X-original-commit: 9b257070027597af704ebb11e3860d2f0d78ac9a
Signed-off-by: Thibault Francois <tfr@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>