[FIX] mrp_mps: indirect demand with starting qty and lead times
Steps:
- create a BOM for product A: component product B, ratio 1.0, lead time 1 day
- put 4 as the on-hand qty of product A
- in the MPS, create records for products A & B with manufacture route for product A
- put 6 as the forecasted demand of product A on the 2nd period or further
(this is so that the indirect demand of B is shown on the previous period)

Issue:
The indirect demand of product B will be 6 instead of 2 (6-4). This is because
it uses `date_stop` with the lead time, see https://github.com/odoo/enterprise/pull/70232

Fix:
Subtract the starting qty from the indirect demand qty for the corresponding subproduct.
This is to make sure that the real `subproduct_indirect_demand` is compared to the
'ratioed' `replenish_qty` of the subproduct when selecting the date to use.

Other issue:
If there's a max replenish qty and lead time, the indirect demand of components is
not distributed.
Fix:
Use the date_start for the demand_qty_dict key.

task 4381021

closes odoo/enterprise#78667

X-original-commit: c156e1e299ea2bbeb9c122fbebdfb4bf42fed8fa
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>