[IMP] account_reports: CTA currency translation: using closing rate instead of current
The way we initially implemented CTA, it was using the current rate for all the balance sheet accounts. Instead, we actually want to consider the closing rate ; that is, the rate of each currency at the end of the fiscal year the operation was made in.

Current rate currency translation is still applied like before when CTA is not enabled on the report.

task-4270102

X-original-commit: 0719c63a646c360a4b090745cd8105128332ef38
Part-of: odoo/enterprise#75602
Related: odoo/odoo#190595
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>