[IMP] pos_iot: certification for scales
In order to comply with the EU Measuring Instruments
Directive, the code for weighing products must be
verified with a checksum to ensure it has not been
modified. This commit adds an endpoint for retrieving
the checksum of the relevant scale code, as well as
refactoring of the code to make it as simpler and
to reduce the amount of code that needs to be certified.

It also adds a status icon to the POS, showing
whether or not it is certified. If it is not
certified, there is an option to auto-fix the
rounding settings to make it certified. This icon
will only show in EU countries.

task-3929839

closes odoo/enterprise#81352

X-original-commit: 49deb0bcfc500a42d535f1a79694ebb15cf385db
Related: odoo/odoo#201582
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>