[IMP] *: gantt: use kanban card in popover
A new attribute "kanban_view_id" is introduced. If set, the kanban card template
of the specified kanban view is used to render the body of the gantt popovers.
If not set, the kanban view (if any) in the gantt view config (e.g. the kanban
view in the curent action) will be used. If no kanban view is found, a default
view based on display_name, date_start, and date_stop (from the gantt arch)
will be generated and used.

Note that this happens only if no content for the popover body has been specified
in a template "gantt-popover" in the gantt arch. In that case the popover body
is constructed using that template.

Kanban view used or not, customization of the popover footer can still be done
via the mechanism introduced in #54292. Note that we cannot use the footer tag
in the kanban card to do that since it has a different usage.

Task ID: 4350860

Part-of: odoo/enterprise#74706
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>