[FIX] {helpdesk_}timesheet{_grid}: make sure trash button is same size than other buttons
Before this commit, the trash button displayed in the wizard used when
the user clicks on stop button in a task (when a timer is running on
that task for that current user) has not the same height than the other
buttons displayed in that wizard.

This commit makes sure the button has the same height than the other
buttons.

task-4382070

closes odoo/enterprise#75512

X-original-commit: 26fab45373d3af2a6ca054820a5d71e2d36023b9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>