[IMP] website_delivery_sendcloud: clean test
Now that the salesperson is set on the cart after its confirmation,
the test failed because it tried to send an email when the mail
configuration was not properly configured.

By relying on the standard test common from website_sale,
we disable the mail sending, and can simplify the existing test
setup.

task-4270127

closes odoo/enterprise#75272

Related: odoo/odoo#184836
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>