[IMP] mrp_workorder: shop floor auto close MO
Currently, in the shop floor, when marking the last WO of an MO as done,
there is the option to also close the MO. This requires users who want
to close the MO to press two buttons in different places.
For users who don't want to close the MO, this can also lead to
confusion whether they marked the WO as done or not.

To improve the user experience, we introduce a new option:
'Close Manufacturing Order' which allows the desired shop floor WO close
behavior to be set in the operation type settings.

When enabled (default), the mark as done button on the last WO of an MO
in the shop floor will have a 'Close Production' button directly instead
of a 'Mark as Done' button.

When disabled, the MO needs to be closed manually after marking the last
workorder as done, as we now no longer show the Close production button
on the workorder after marking it as done.

task-4367540

closes odoo/enterprise#75307

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>