[IMP] industry_fsm: log note to notify timer actions in fsm task
Before this commit, when the current user starts and stops a timer on a
fsm task, a log note is logged in the chatter of that task to say when
the timer has been started and stopped.

This commit improve a bit that behavior since now the timesheet could be
started inside the timesheets app. If the timer is started inside the
timesheet app and the task set when the timer is started is a fsm task
then the log note will be logged in the task. When the timer is stopped
in timesheets app then if the task linked to the timesheet with the
timer running is a fsm one then we will log a note in that fsm task as
well.

task-2276015