[FIX] documents: fix the copy links feature in documents list view
In list view, the "Copy links" action was returning empty string (or undefined),
since we haven't added the field `access_url` to the view, so accessing it on
the `record.data` from JS returned `undefined`.

Now, we add `access_url` field as a hidden field in the list view.

We also add a fallback method in JS `populateMissingAccessURLsIfAny`
to fetch the `access_url` for clients who don't update the XML file.

opw-4309365

closes odoo/enterprise#75210

X-original-commit: f0dddb3c74a2c53936911034f3cdc01e2533e431
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Hadi El Yakhni (hael) <hael@odoo.com>