[FIX] account_reports: Fixed tax report default date selector
This commit makes sure that, in tax report, when tax_period is used to determine the report date,
and when the tax_period periodicity matches one of the 3 standard filters (month, quarter, or year),
that the correponding standard filter is selected with the right date set on it.

Before this commit, the default date filter of the tax report would make use of the tax_period to set its date.
The date was set properly, but the tax_period date filter would be hidden in case the periodicity matches monthly, quarterly, or annually.
When its hidden, the date filter dropdown menu would show 3 standard filters with none of them selected, which was confusing.

task-4373943

closes odoo/enterprise#77504

X-original-commit: 22878ec7e38c2a9fbf871beb239d01a42a17546a
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>