[FIX] l10n_uk_reports: Missing fields in the fraud prevention headers
Before when sending request to hmrc we had a fraud prevention headers creation but only with part of the data needed, client_data was provided for one of the two request.
Another problem was the ips. If we are in a private network we should still send the public IP.

task-4387562

closes odoo/enterprise#79204

X-original-commit: 6c0880bfd372198dfae822370ef1d51bd6b6d607
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>