[FIX] l10n_kr_reports: report codes
Some code used in accounting report lines are
also used in other reports outside of the KR
localization.
As of now, unlike what the constrains hint, codes
are not expected to be unique per report but instead
unique entirely which was not the case.
The Unallocated Earnings lines had the same code
as the standard ones which ended up causing some
tests to fail.

To fix this, we are adding a prefix to all codes.
The constrains was also temporarily changed to ensure
that the new codes are at least unique in the KR module
and do not clash with account_reports.

closes odoo/enterprise#75419

X-original-commit: 4e20f051b848d00f0a25fb98fbf1232b1e31f72d
Related: odoo/odoo#190204
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>