[IMP] timer: manage value field name prop in timer start field
Before this commit, it was possible to display the time + the
unit_amount value of the timesheet with the timer running since the
unit_amount could be higher than 0 when the timer starts on that
timesheet, meaning the timer displayed in the task/ticket should take
into account the unit amount set on the timesheet.

This commit defines a new option in timer start field widget to be able
to add the field containing the unit_amount of the timesheet with the
timer running in that task/ticket to correctly take into account that
unit amount when the timer is displayed in task/ticket.

task-2276015