[IMP] account_accountant: intensify hard lock date UX warning
This commit improves the UX design of the lock date wizard, by
intensifying the warning on the hard lock date field, showing them in an
`alert-danger`.

We want to make it even more explicit that the hard lock is
irreversible, concerns all users, given we already have feedbacks from
users who messed with it by setting it too early.

closes odoo/enterprise#78919

Task-id: 4391192
X-original-commit: ac67ebccbb06e764ac60c3f145e0423c80b189e6
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>