[IMP] iot: do not send handlers up to date
We currently download handlers from db on every restart of odoo service, even if
they are up to date.

We now compute the sha256 hash of the handlers, compare it with the one the IoT Box
saved the first time it downloaded it and return a 304 (Not Modified) if
they are already the latest version.

We now also avoid sending handlers not corresponding to the platform (we don't
send Windows drivers to physical IoT Box or Linux drivers to Windows IoT Box
anymore).

Community PR: [https://github.com/odoo/odoo/pull/190245](https://github.com/odoo/odoo/pull/190245)

closes odoo/enterprise#75443

Task: 4107538
Related: odoo/odoo#190245
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>