[FIX] l10n_be_soda: depends on accountant
Steps to reproduce:
- Install account_accountant & l10n_be
- l10n_be_soda is auto installed
- Modify a bank transaction OR
- Try to configure SODA from settings

Access Error:
The write method on account.move.line, and the SODA import wizard require `group_account_user` access. But this group is only activated with accounting.

This module depends on Invoicing (account_accountant) but it should depend on Accounting (accountant).
Since this change is not possible for stable, we check for access to the SODA models (`group_account_user`) while the correct fix is applied in master 18.1 (https://github.com/odoo/enterprise/pull/75505)

Issue reported by AVW

closes odoo/enterprise#75574

X-original-commit: 743d63d66a1d33e8ab4156a4b35f6f5da66db299
Signed-off-by: William André (wan) <wan@odoo.com>