[FIX] spreadsheet_edition: allow base user to restore a spreadsheet
Base users do not have access to the model `spreadsheet.revision` which
is directly affected in a collaborative context. The usual flow consists
to give access to the revisions once we check the access rights on the
related spreadsheet. The code that handled the restoration of previous
versions did not follow this flow and base users would face access
errors.

Task-4397512

closes odoo/enterprise#76601

X-original-commit: 89d1bee2b30c3317374f6de0a6221e1fca63667b
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>