[FIX] documents_spreadsheet: fix warning when inputing a negative filter offset
In unit tests, when global filter offsets are added with negative values, we have
a warning in the console saying that the symbol '-' cannot be parsed. The reason
behind the warning is that the input element triggers an onChange event whenever
a new character is added. The commit fixes this issue by setting the option `instantly`
to true to only trigger the onChange with the whole new value.
task-4397783

closes odoo/enterprise#75477

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>