[MERGE] l10n_ke_hr_payroll_shif: NHIF to SHIF
Problem
----------
Since 1st October, the SHIF replace the NHIF, the only difference is the computation of the amount.
All payslip before and on the 9 October use the NHIF and all new payslip from the 9 October use the SHIF

Objective
----------
Adapt the salary rules to compute the good rule.
Duplicate reports and wizard to make a smooth transition and keep both of them

Solution
----------
Create new SHIF parameter with a min amount and rate computation.
Duplicate all occurrences of NHIF and rename them SHIF (for the reports)

task-4294419

Part-of: odoo/enterprise#73456
Related: odoo/upgrade#6868
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>