[IMP] *expense*: Bye Bye reports
Most of the time, expenses are created as "standalone"s and forcing
the use of an expense report is adding frustrating unnecessary steps.

A simplification was decided and the model hr.expense.sheet is removed,
giving all of its remaining important logic to the `hr.expense`.

In addition, approval flow was simplified with the implementation of
an auto-validation process, where submitting an expense also approves it
when no expense manager is set on the employee.

See task for more details.

closes odoo/enterprise#75507

Task-id: 4481615
Related: odoo/odoo#189701
Related: odoo/upgrade#6990
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>