[FIX] l10n_be_soda: depends on accountant
Steps to reproduce:
- Install account_accountant & l10n_be
- l10n_be_soda is auto installed
- Modify a bank transaction OR
- Try to configure SODA from settings

Access Error:
The write method on account.move.line, and the SODA import wizard require `group_account_user` access. But this group is only activated with accounting.

This module currently depends on Invoicing (account_accountant) but it should depend on Accounting (accountant).
Furthermore, It doesn't make sense to import journal entries if accounting is not available.

Issue reported by AVW

closes odoo/enterprise#75505

Signed-off-by: William André (wan) <wan@odoo.com>