[FIX] stock_barcode_mrp: block product addition if 'Allow Extra Product' is off
Issue Before This Commit
========================
The barcode application for the 'Manufacturing' operation type allowed
products to be scan and added to a Manufacturing Order (MO) even when
the 'Allow Extra Product' setting was disabled.

Steps to Produce:
=================
- Disable the 'Allow Extra Product' option for the 'Manufacturing'
  operation type.
- Create a new MO or open an existing MO.
- Scan a product other than the MO product still allowed it to be added,
  despite the setting being disabled.

This commit removes the code that allowed extra products to be added,
ensuring compliance with the operation type setting.

closes odoo/enterprise#75535

Task: 4369998
Signed-off-by: Steve Van Essche <svs@odoo.com>