[FIX] hr_payroll: comment displayed line
**Issue:**
If a user changes a salary rule to the default Python Code for computation, an error occurs when trying to `Compute Sheet` on an employee's payslip.

**Expected:**
The computation should be feasible.

**Steps to reproduce:**
- Activate Payroll app;
- Move to `Configuration/Salary:Rules`;
- Select any rule having its `Computation` `Amount Type` field not set to `Python Code` (e.g. `Regular Pay > House Rent Allowance`);
- Change the `Amount Type` to `Python Code`;
- Create a new payslip entry for any employee and change the structure for the one you changed the rule in (e.g. `Regular Pay`);
- Click `Compute Sheet`;

**Cause:**
The Python code displayed in the computation rule contains an uncommented comment line that.

**Fix:**
Comment the blocking line.

opw-4377712

closes odoo/enterprise#75558

X-original-commit: b2616fa12844501c65dfaac49a1ccefd53512308
Signed-off-by: Pierre Lamotte (pila) <pila@odoo.com>