[IMP] documents: allow portal user to delete their own documents
Instead of enabling the Trash for the portal user, we add the delete action for
their own documents only so that a user that have uploaded a document by
mistake can delete it. Indeed, as portal user don't have access to the Trash,
moving a document to the Trash like an internal user would do is no use.

Task-4221258

Part-of: odoo/enterprise#77645
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>