[FIX] documents_hr_payroll: fix wrong contact
How to reproduce:
- Install documents_hr_payroll
- Go to Marc Demo employee file
- Activate debug and click on View:form item of the debug menu
- Move the field "work_contact_id" in the sheet tag and make it visible
- Return on the Marc Demo employee file
- Set a new work contact for Marc Demo (using the field): newContact
- Go to the payslip and confirm one of the payslip of Marc
- Open HR Payroll directory in the Document App
- The payslip of Marc is present, select it and click on "Share"

The payslip is shared with "Mitchell Admin" and "newContact" so Marc Demo won't
be able to see it because it is not shared with the partner of his user.

The problem was that the employee work_contact_id partner was used to give
access to the payslip (and as contact of the payslip document) while for some
employee, the work contact is different than the partner associated with their
user. In that case, the payslip is shared with a partner different than
the employee user partner and the employee cannot see the payslip.

We solve the problem by always sharing the payslip with the partner of the
employee user if it exists and we fallback only to the work contact if there is
no user associated with the employee.

Task-4404015

closes odoo/enterprise#77288

X-original-commit: d408e3182f4c7a5477c05ecf09f29a603f6f6a40
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>