[IMP] stock_barcode: operation type kanban card colour adjustments
The cards in the barcode kanban view for picking operation types
should have their standard colour streak on the left border instead of
the entire card background. A colour picker for the `color` field was
also added to the operation type form view to allow the user to adjust
it as needed.

Task ID: 4385345

closes odoo/enterprise#75551

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>