[FIX] social_twitter: reenable likes from the backend
This commit fixes an issue with the social kanban view when liking a Twitter/X
post from it.

Since odoo/odoo#189109, some views encountered issues with disappearing props
and various crashes.
The social kanban view used the `list` props to update the like counters on
potentially several posts. Since this one disappeared it needed to be replaced
to be functional again.
This is what this commit does: instead of searching records on `this.props.list`
the method does it on `this.props.group.model.root`.

task-4403372

closes odoo/enterprise#75555

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>