[FIX] web_gantt: pill resize invalid operation
This commit fixes an issue with pill resizing: if the result of the
resize would have the start date after the end date or the end date
before the start date, it would result in an invalid operation error.
After this commit, such an operation will only result in being cancelled
with a warning notification to explain the problem.

task-4320664

closes odoo/enterprise#75674

X-original-commit: 833bff2da985c278ebc379cd48bd602c489f3846
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>