[IMP] account_reports: properly handle the unfold of groupby lines depending on the existence of their sublines
For new custom reports, we now wish to be able to generate groupby values computed from actual move lines that are not considered as having sublines, and are hence never unfoldable to display further grouping levels (typically for the initial balance column in the Trial Balance).

To do that, we tried relying on the has_sublines key returned in the results of the engines, but this didn't work well, and revealed the fact the existence of sublines was aggregated by expression (instead of being kept by grouping key), and was actually never called when deciding whether a groupby line should be unfoldable or not.

Instead, the existence of additional levels of grouping was used ; assuming that an existing groupby line generated for level N would also generate at least one line at level N+1. This criteria hid the problem so far, and made everything work fine, but we now need to get rid of it for the new use cases we have to support.

Part-of: odoo/enterprise#73298
Related: odoo/odoo#186255
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>