[FIX] delivery_shiprocket: delivery label total value
Issue Before This Commit:
--------------------------------
The delivery label generated by the Shiprocket delivery method displayed an
incorrect total when a discount or free delivery was applied,
which caused confusion for users.

Steps to Produce:
--------------------------------
1. Apply a discount or free delivery for a Shiprocket shipment.
2. Validate the delivery.
3. The total on the delivery label will show an incorrect value,
   not accounting for the discount or free delivery.

With This Commit:
--------------------------------
The issue is fixed by ensuring the correct data is sent to the API,
which now reflects the accurate total value on the delivery label,
even when a discount or free delivery is applied.

closes odoo/enterprise#77147

Task-id: 4395054
X-original-commit: a931774e8221fbdb815976783513225ce456bccd
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Sahil Koladiya (sako) <sako@odoo.com>