[IMP] l10n_au_hr_payroll: Add income stream type on opening balances
All year to date values are expected to be reported
per income stream type. That requires all the opening
balance to be imported with an income stream type
if it has changed prior to transferring to Odoo. This
commit adds the income stream type field to the opening
balances and the import flow.

closes odoo/enterprise#75693

Related: odoo/upgrade#6957
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>