[IMP] industry_fsm: add domain to Tasks analysis report in Field Service
Before this commit, due to the changes made in community, the all
subtasks will be included in the main view of tasks or Tasks analysis in
Project app when the filter `Show tasks` will be enabled in the search
view.

This commit adds a domain to only see the tasks displayed in the main
views of tasks as it is the case in tasks analysis report in Project
app.

task-4295713

closes odoo/enterprise#73124

Related: odoo/odoo#185912
Related: odoo/upgrade#6811
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>