[FIX] sale_subscription: prevent AccessError in kanban
Steps to reproduce:
-------------------

	- Install sale_subscription and inventory.
	- Create a subscription product.
	- Log in as a non-Sales user (e.g., Inventory user).
	- Access the product list in Kanban view.(from stock)

Issue:
-------

Non-Sales users encounter an AccessError when the Kanban view
attempts to read product_subscription_pricing_ids.
This is due to:
 - "record.product_subscription_pricing_ids.raw_value.length"
Attempted to be computed even if the user does not have the access rights.

Fix:
-----

Updated the Kanban view to conditionally render subscription pricing fields
only for users in the Sales group.

opw-4404190

closes odoo/enterprise#76571

X-original-commit: d0d09acfe898d8b6fe4187edbb775ea26a9741b8
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Julien Schwanen (scju) <scju@odoo.com>