[FIX] account_accountant: 'Invalid Statement' filter condition
--Before--

The 'Invalid Statements' filter only considered the case of the Ending Balance not matching the Starting Balance + the sum of its transactions. We were not considering the case of the Starting Balance of the statement not matching the previous statement's Ending Balance. 

--Now--

Fixed the condition of the filter to account for the second case.

task-4397412

closes odoo/enterprise#76067

X-original-commit: 152fa55af4a9392de57c3d709b475a7478edbe83
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>