[IMP] account_accountant: allow reconciliation of draft entries
Since accountants from an external firm typically review client files only once per VAT period (every three months),
it's a pain to wait for them to confirm the bills before reconciling. To avoid this, and to simplify the process for
users who might pay on draft bills, we allow the reconciliation of draft entries.

That reconciliation with a draft entry might trigger an automatic entry, such as an exchange or cash-basis move, that
is created and left in draft as well. The posting of if would be delayed until of related journal entries are posted.
Note that in case the draft entry changed between the moment we reconciled it and the moment we post it, in a way that
the exchange or cash-basis move would have been different, we simply break the reconcilation and the user will have to
redo it.

Task-4286245

closes odoo/enterprise#75385

Related: odoo/odoo#190110
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Co-authored-by: h4818 <ayh@odoo.com>