[IMP] timesheet_grid, helpdesk_timesheet: stop timer when timesheets disabled
Before this commit, when the user goes to a task and starts a timer,
then changes the project of the task to put a project without the
timesheets feature enabled. The timer is still running on the task even
if the feature is not available on the new project.

This commit makes sure the timer is removed if the project set on the
task does not have the feature enabled. It also applies the same logic
on helpdesk tickets if the new team does not have the timesheet feature
enabled.

task-2276015